+++ src/main/java/org/openlmis/cce/util/Resource2Db.java
@@ -5,12 +5,12 @@
  * This program is free software: you can redistribute it and/or modify it under the terms
  * of the GNU Affero General Public License as published by the Free Software Foundation, either
  * version 3 of the License, or (at your option) any later version.
- *  
+ *
  * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  * See the GNU Affero General Public License for more details. You should have received a copy of
  * the GNU Affero General Public License along with this program. If not, see
- * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
+ * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org.
  */
 
 package org.openlmis.cce.util;
@@ -22,6 +22,8 @@ import org.apache.commons.collections4.IteratorUtils;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVRecord;
+import org.apache.commons.io.ByteOrderMark;
+import org.apache.commons.io.input.BOMInputStream;
 import org.apache.commons.lang3.Validate;
 import org.apache.commons.lang3.tuple.MutablePair;
 import org.apache.commons.lang3.tuple.Pair;
@@ -110,12 +112,13 @@ public class Resource2Db {
    and Pair.right is the rows of data which go into those columns (each row is an array, the array
    matches the order of the columns
    */
-  private Pair<List<String>, List<Object[]>> resourceCsvToBatchedPair(final Resource resource)
+  Pair<List<String>, List<Object[]>> resourceCsvToBatchedPair(final Resource resource)
       throws IOException {
     XLOGGER.entry(resource.getDescription());
 
     // parse CSV
-    try (InputStreamReader isReader = new InputStreamReader(resource.getInputStream())) {
+    try (InputStreamReader isReader = new InputStreamReader(
+        new BOMInputStream(resource.getInputStream(), ByteOrderMark.UTF_8))) {
       CSVParser parser = CSVFormat.DEFAULT.withHeader().withNullString("").parse(isReader);
 
       // read header row
@@ -143,7 +146,7 @@ public class Resource2Db {
   /*
    runs the list of SQL strings directly on the database - could be insert / update
    */
-  private void updateDbFromSqlStrings(final List<String> sqlLines) {
+  void updateDbFromSqlStrings(final List<String> sqlLines) {
     XLOGGER.entry();
 
     if (CollectionUtils.isEmpty(sqlLines)) {