+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 Bug fixes:
 * [OLMIS-4588](https://openlmis.atlassian.net/browse/OLMIS-4588): Fixed incorrect link in the CCE notification
 
+Improvements:
+* [OLMIS-4647](https://openlmis.atlassian.net/browse/OLMIS-4647): Added Jenkinsfile
+
 1.0.0 / 2018-04-24
 ==================