+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Bug fixes:
 Improvements:
 * [OLMIS-4647](https://openlmis.atlassian.net/browse/OLMIS-4647): Added Jenkinsfile
 * [OLMIS-4905](https://openlmis.atlassian.net/browse/OLMIS-4905): Updated notification service to use v2 endpoint.
+* [OLMIS-4876](https://openlmis.atlassian.net/browse/OLMIS-4876): Applied new demo data loading approach
 
 1.0.0 / 2018-04-24
 ==================